+++ src/components/scenes/ChangePasswordScene.tsx
@@ -1,7 +1,7 @@
 import { EdgeAccount } from 'edge-core-js'
 import * as React from 'react'
 import { Keyboard, Platform, ScrollView } from 'react-native'
-import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scroll-view'
+import { KeyboardAwareScrollView } from 'react-native-keyboard-controller'
 import { cacheStyles } from 'react-native-patina'
 
 import { lstrings } from '../../common/locales/strings'
@@ -196,9 +196,7 @@ const ChangePasswordSceneComponent = ({
         <KeyboardAwareScrollView
           contentContainerStyle={[styles.container, keyboardPadding]}
           keyboardShouldPersistTaps="handled"
-          extraScrollHeight={theme.rem(6)}
-          enableAutomaticScroll
-          enableOnAndroid
+          bottomOffset={theme.rem(6)}
         >
           {content}
         </KeyboardAwareScrollView>