+++ src/components/scenes/newAccount/NewAccountUsernameScene.tsx
@@ -1,6 +1,6 @@
 import { asMaybeNetworkError, EdgeAccount } from 'edge-core-js'
 import * as React from 'react'
-import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scroll-view'
+import { KeyboardAwareScrollView } from 'react-native-keyboard-controller'
 import { cacheStyles } from 'react-native-patina'
 import { sprintf } from 'sprintf-js'